# truck-backend/openapi.yaml
swagger: "2.0"
info:
  description: "Homework description"
  version: "1.0.0"
  title: "LAB FE NodeJS HW #3"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "Kyrylo_Yezhov@epam.com"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /api/loads:
    get:
      tags:
      - "Load"
      summary: "Get user's loads"
      description: "Retrieve the list of loads for authorized user, returns list of completed and active loads for Driver and list of all available loads for Shipper"
      operationId: "getUserLoads"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "query"
          name: "status"
          description: "Filter by load status"
          required: false
          type: string
        - in: "query"
          name: "limit"
          description: "Limit for records per request, pagination parameter(default: 10, max: 50)"
          required: false
          type: integer
        - in: "query"
          name: "offset"
          description: "Offset for records, pagination parameter(default: 0)"
          required: false
          type: integer
      responses:
        "200":
          description: "Success"
          schema:
            type: "object"
            properties: 
                loads:
                    type: array
                    items:
                        $ref: "#/definitions/Load"
        "400":
          description: "Bad request"
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: "Internal server error"
          schema:
            $ref: '#/definitions/Error'
      security:
      - jwt_token: []
    post:
        tags:
        - "Load"
        summary: "Add Load for User"
        description: "Add Load for User(abailable only for shipper role)"
        operationId: "addUserLoad"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - in: "body"
          name: "Load payload"
          schema:
            type: "object"
            properties:
              name:
                type: "string"
                example: "Moving sofa"
              payload:
                type: "number"
                example: 100
              pickup_address:
                type: "string"
                example: "Flat 25, 12/F, Acacia Building 150 Kennedy Road"
              delivery_address:
                type: "string"
                example: "Sr. Rodrigo Domínguez Av. Bellavista N° 185"
              dimensions:
                type: "object"
                properties: 
                  width:
                    type: "number"
                    example: 44
                  length:
                    type: "number"
                    example: 32
                  height:
                    type: "number"
                    example: 66
        responses:
            "200":
                description: "Success"
                schema:
                    type: "object"
                    properties:
                      message: 
                        type: string
                        example: "Load created successfully"
            "400":
                description: "Bad request"
                schema:
                    $ref: '#/definitions/Error'
            "500":
                description: "Internal server error"
                schema:
                    $ref: '#/definitions/Error'
        security:
        - jwt_token: []
...
  /api/loads/{id}:
    get:
        tags:
        - "Load"
        summary: "Get user's Load by id"
        description: "Get user's Load by id"
        operationId: "getUserLoadById"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - in: "path"
          name: "id"
          description: "Load id"
          required: true
          type: string
        responses:
            "200":
              description: "Success"
              schema:
                type: object
                properties:
                  load:
                      $ref: '#/definitions/Load'
            "400":
              description: "Bad request"
              schema:
                  $ref: '#/definitions/Error'
            "500":
              description: "Internal server error"
              schema:
                  $ref: '#/definitions/Error'
        security:
        - jwt_token: []
    put:
        tags:
        - "Load"
        summary: "Update user's load by id"
        description: "Update user's load by id (abailable only for shipper role)"
        operationId: "updateUserLoadById"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - in: "path"
          name: "id"
          description: "Load id"
          required: true
          type: string
        - in: "body"
          name: "Load payload"
          schema:
            type: "object"
            properties:
              name:
                type: "string"
                example: "Moving sofa"
              payload:
                type: "number"
                example: 100
              pickup_address:
                type: "string"
                example: "Flat 25, 12/F, Acacia Building 150 Kennedy Road"
              delivery_address:
                type: "string"
                example: "Sr. Rodrigo Domínguez Av. Bellavista N° 185"
              dimensions:
                type: "object"
                properties: 
                  width:
                    type: "number"
                    example: 44
                  length:
                    type: "number"
                    example: 32
                  height:
                    type: "number"
                    example: 66
        responses:
            "200":
                description: "Success"
                schema:
                    type: "object"
                    properties:
                      message: 
                        type: string
                        example: "Load details changed successfully"
            "400":
              description: "Bad request"
              schema:
                  $ref: '#/definitions/Error'
            "500":
              description: "Internal server error"
              schema:
                  $ref: '#/definitions/Error'
        security:
        - jwt_token: []
    delete:
        tags:
        - "Load"
        summary: "Delete user's load by id"
        description: "Delete user's load by id (abailable only for shipper role)"
        operationId: "deleteUserLoadById"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - in: "path"
          name: "id"
          description: "Load id"
          required: true
          type: string
        responses:
            "200":
                description: "Success"
                schema:
                    type: "object"
                    properties:
                      message: 
                        type: string
                        example: "Load deleted successfully"
            "400":
              description: "Bad request"
              schema:
                  $ref: '#/definitions/Error'
            "500":
              description: "Internal server error"
              schema:
                  $ref: '#/definitions/Error'
        security:
        - jwt_token: []
...
securityDefinitions:
  jwt_token:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  Error:
    type: "object"
    properties:
      message:
        type: "string"
        description: "Error message"
...
  Load:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "5099803df3f4948bd2f98391"
      created_by:
        type: "string"
        example: "5099803df3f4948bd2f98391"
      assigned_to:
        type: "string"
        example: "5099803df3f4948bd2f98391"
      status:
        type: "string"
        enum: ["NEW", "POSTED", "ASSIGNED", "SHIPPED"]
        example: "NEW"
      state:
        type: "string"
        enum: ["En route to Pick Up", "Arrived to Pick Up", "En route to delivery", "Arrived to delivery"]
        example: "En route to Pick Up"
      name:
        type: "string"
        example: "Moving sofa"
      payload:
        type: "number"
        example: 100
      pickup_address:
        type: "string"
        example: "Flat 25, 12/F, Acacia Building 150 Kennedy Road"
      delivery_address:
        type: "string"
        example: "Sr. Rodrigo Domínguez Av. Bellavista N° 185"
      dimensions:
        type: "object"
        properties: 
          width:
            type: "number"
            example: 44
          length:
            type: "number"
            example: 32
          height:
            type: "number"
            example: 66
      logs:
        type: "array"
        items: 
          type: "object"
          properties:
            message:
              type: "string"
              example: "Load assigned to driver with id ###"
            time:
              type: "string"
              format: "date"
              example: "2020-10-28T08:03:19.814Z"
      created_date:
        type: "string"
        format: "date"
        example: "2020-10-28T08:03:19.814Z"
...